Moved applications endpoints to a separate file
